@@ -1,58 +1,47 @@
 {
   "name": "windowbuilder-reports",
   "description": "Reports and image rendering for windowbuilder project",
-  "version": "0.1.0",
+  "version": "2.0.233",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
   "license": "SEE LICENSE IN LICENSE.en.md",
   "main": "index.js",
   "engines": {
-    "node": "^7.10",
-    "npm": "^5.0"
+    "node": "^8.8",
+    "npm": "^5.5"
   },
   "scripts": {
-    "start": "node ./index",
+    "start": "node --inspect-brk ./index",
+    "build": "node ./scripts/build",
     "canvas": "apt-get install libcairo2-dev libjpeg62-turbo-dev libpango1.0-dev libgif-dev build-essential g++",
-    "meta:prebuild": "node ./scripts/prebuild.meta.js",
-    "meta:link": "npm link metadata-core metadata-redux metadata-pouchdb metadata-abstract-ui metadata-react"
+    "meta:prebuild": "node ./scripts/meta.prebuild",
+    "meta:link": "node scripts/meta.link"
   },
   "dependencies": {
-    "alasql": "^0.4.1",
-    "koa": "^2.3.0",
+    "@koa/cors": "^2.2.1",
+    "koa": "^2.4.1",
     "koa-better-router": "^2.1",
-    "metadata-abstract-ui": "^2.0.2-beta.25",
-    "metadata-core": "^2.0.2-beta.25",
-    "metadata-pouchdb": "^2.0.2-beta.25",
-    "moment": "^2.18",
-    "paper-jsdom-canvas": "^0.11.4",
-    "pouchdb": "^6.3.4",
-    "pouchdb-adapter-http": "^6.3.4",
-    "pouchdb-adapter-idb": "^6.3.4",
-    "pouchdb-adapter-memory": "^6.3.4",
-    "pouchdb-adapter-websql": "^6.3.4",
-    "pouchdb-authentication": "^0.5",
-    "pouchdb-core": "^6.3.4",
-    "pouchdb-find": "^6.3.4",
-    "pouchdb-mapreduce": "^6.3.4",
-    "pouchdb-replication": "^6.3.4"
+    "md5-file": "^3.2.3",
+    "metadata-abstract-ui": "^2.0.16-beta.46",
+    "metadata-core": "^2.0.16-beta.46",
+    "metadata-pouchdb": "^2.0.16-beta.46",
+    "paper-jsdom-canvas": "^0.11.5"
   },
   "devDependencies": {
-    "clipboard": "^1.7",
-    "debug": "^3.0.0",
-    "eslint": "^4.4.1",
-    "eslint-config-standard": "^10.2",
-    "eslint-config-standard-react": "^5.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "eslint-plugin-import": "^2.5",
-    "eslint-plugin-node": "^5.1.1",
-    "eslint-plugin-promise": "^3.5",
-    "eslint-plugin-react": "^7.2.0",
-    "eslint-plugin-standard": "^3.0",
+    "eslint": "^4.15.0",
+    "eslint-config-react-app": "^2.1.0",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.5.1",
     "gulp": "^3.9",
     "gulp-concat": "^2.6",
     "gulp-rename": "^1.2",
-    "gulp-replace": "^0.6",
-    "gulp-strip-comments": "^2.5.1",
-    "gulp-wrap": "^0.13.0"
+    "gulp-replace": "^0.6.1",
+    "gulp-strip-comments": "^2.5.2",
+    "gulp-wrap": "^0.13.0",
+    "rollup": "^0.51",
+    "rollup-plugin-cleanup": "^2.0",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-resolve": "^3.0.2",
+    "rollup-plugin-replace": "^2.0.0"
   },
   "homepage": "http://www.oknosoft.ru/zd/",
   "repository": {
